refactoring such pages: test cases for start, main, create post pages
