+++ constants/header.py
@@ -1,0 +1,7 @@
+class HeaderConstants:
+    LOGO_LINK_XPATH = ".//h4"
+    SEARCH_ICON_XPATH = ".//*[@data-original-title='Search']"
+    CHAT_ICON_XPATH = ".//*[@data-original-title='Chat']"
+    MY_PROFILE_ICON_XPATH = ".//*[@data-original-title='My Profile']"
+    CREATE_POST_BUTTON_XPATH = ".//*[text()='Create Post']"
+    SIGN_OUT_BUTTON_XPATH = ".//button[text()='Sign Out']"
