+++ constants/post_page.py
@@ -1,0 +1,5 @@
+class PostPageConstants:
+    SUCCESS_MESSAGE_TEXT = 'New post successfully created.'
+    MESSAGE_NEW_POST_CREATED_XPATH = ".//*[text()='New post successfully created.']"
+    TITLE_XPATH = ".//h2"
+    REMOVE_ICON_XPATH = ".//button[@data-toggle='tooltip']"
