+++ pages/main_page.py
@@ -1,7 +1,6 @@
 from constants.main_page import MainPageConstants
 from pages.base import BasePage
-from pages.create_post import CreatePost
-from pages.my_profile import MyProfile
+from pages.header import Header
 from pages.utils import log_decorator
 
 
@@ -10,6 +9,7 @@ class MainPage(BasePage):
     def __init__(self, driver):
         super().__init__(driver)
         self.constants = MainPageConstants()
+        self.header = Header(self.driver)
 
     @log_decorator
     def verify_welcome_message(self, username):
@@ -18,44 +18,14 @@ class MainPage(BasePage):
         assert hello_user.text == f"Hello {username.lower()}, your feed is empty."
 
     @log_decorator
-    def logout(self):
-        """Log out from user account"""
-        self.wait_until_element_enabled(value=self.constants.SIGN_OUT_BUTTON_XPATH).click()
-
-    @log_decorator
-    def refresh_main_page(self):
-        """Refresh page by tapping logo-link"""
-        self.wait_until_element_enabled(value=self.constants.LOGO_LINK_XPATH).click()
-
-    @log_decorator
-    def transition_to_search_bar(self):
-        """Open search bar"""
-        self.wait_until_element_enabled(value=self.constants.SEARCH_ICON_XPATH).click()
-
-    @log_decorator
     def verify_search_bar_opened(self):
         """Verify search bar is opened"""
         search_bar = self.wait_until_find_element(value=self.constants.PLACEHOLDER_SEARCH_BAR_XPATH)
         assert search_bar.is_displayed()
 
     @log_decorator
-    def transition_to_chat_form(self):
-        """Open Chat form"""
-        self.wait_until_element_enabled(value=self.constants.CHAT_ICON_XPATH).click()
-
-    @log_decorator
     def verify_chat_form_opened(self):
         chat_form = self.wait_until_find_element(value=self.constants.CHAT_FORM_XPATH)
         assert chat_form.text == "Chat"
 
-    @log_decorator
-    def transition_to_my_profile(self):
-        """Open my profile page"""
-        self.wait_until_find_element(value=self.constants.MY_PROFILE_ICON_XPATH).click()
-        return MyProfile(self.driver)
 
-    @log_decorator
-    def transition_to_create_post_page(self):
-        """Open Create post page"""
-        self.wait_until_find_element(value=self.constants.CREATE_POST_BUTTON_XPATH).click()
-        return CreatePost(self.driver)
