+++ tests/test_cases_start_page.py
@@ -1,35 +1,12 @@
 """Stores tests related to Start page and Main Page"""
 
 import pytest
-from selenium.webdriver.chrome import webdriver
 
 from conftest import BaseTest
-from constants.base import BaseConstants
-from pages.start_page import StartPage
 from pages.utils import User
 
 
 class TestStartPage(BaseTest):
-
-    @pytest.fixture(scope='function')
-    def driver(self):
-        """Create and return driver, close after test"""
-        driver = webdriver.WebDriver(BaseConstants.DRIVER_PATH)
-        yield driver
-        driver.close()
-
-    @pytest.fixture(scope="function")
-    def random_user(self):
-        """Creates values object for random user"""
-        user = User()
-        user.fill_properties()
-        return user
-
-    @pytest.fixture(scope="function")
-    def start_page(self, driver):
-        """Return start page object"""
-        driver.get(BaseConstants.URL)
-        return StartPage(driver)
 
     @pytest.fixture(scope="function")
     def registered_user(self, start_page, random_user):
@@ -37,7 +14,7 @@ class TestStartPage(BaseTest):
         # Fill fields with provided  values
         main_page = start_page.register_user(random_user)
         # Logout
-        main_page.logout()
+        main_page.header.logout()
         return random_user
 
     def test_invalid_login(self, start_page, random_user):
